# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  aslon1213_gift_pkg_repository.Empty:
    type: object
...
  aslon1213_gift_pkg_repository.Income:
    properties:
      amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      source:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  aslon1213_gift_pkg_repository.Response-array_aslon1213_gift_pkg_repository_Income:
    properties:
      data:
        items:
          $ref: '#/definitions/aslon1213_gift_pkg_repository.Income'
        type: array
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
...
  aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty:
    properties:
      data:
        $ref: '#/definitions/aslon1213_gift_pkg_repository.Empty'
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
...
  aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Income:
    properties:
      data:
        $ref: '#/definitions/aslon1213_gift_pkg_repository.Income'
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: hamidovaslon1@gmail.com
    name: API Support
    url: https://github.com/aslon1213/gift
  description: This is a Gift API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/aslon1213/gift
  title: Gift API
  version: 0.1.0
paths:
...
  /api/v1/incomes:
    get:
      description: Returns all incomes for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-array_aslon1213_gift_pkg_repository_Income'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: List incomes
      tags:
      - incomes
    post:
      consumes:
      - application/json
      description: Creates a new income entry for the authenticated user
      parameters:
      - description: Income data
        in: body
        name: income
        required: true
        schema:
          $ref: '#/definitions/aslon1213_gift_pkg_repository.Income'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Income'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: Create income
      tags:
      - incomes
  /api/v1/incomes/{id}:
    delete:
      description: Deletes an income entry for the authenticated user by its ObjectID
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: Delete income
      tags:
      - incomes
    get:
      description: Returns the income by its ObjectID, for the authenticated user
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Income'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: Get income by ID
      tags:
      - incomes
    put:
      consumes:
      - application/json
      description: Updates an existing income by its ObjectID for the authenticated
        user
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: string
      - description: Income data
        in: body
        name: income
        required: true
        schema:
          $ref: '#/definitions/aslon1213_gift_pkg_repository.Income'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Income'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: Update income
      tags:
      - incomes
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"